Refactor GitHub Actions workflow file to use latest versions and improve Docker setup
Update the GitHub Actions workflow file to use the latest versions of the actions/checkout and docker/setup-buildx-action actions. This ensures that the code is checked out correctly and Docker Buildx is set up properly.

Also, update the image tag in the deployment.yaml file to use the latest version of the ollama-ubi9 image from the quay.io/takinosh repository.

This refactor improves the reliability and efficiency of the workflow and ensures that the correct image version is used in the deployment.